requirements.txt
discord.py
APScheduler
pymysql
python-dotenv
asyncio
aiohttp